Move IncomingMessagesProcessor to Application
