+++ src/Win2Mqtt.Application/ServiceCollectionExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
+using Win2Mqtt.SystemActions;
 using Win2Mqtt.SystemMetrics;
 
 namespace Win2Mqtt.Application
@@ -7,6 +8,7 @@ namespace Win2Mqtt.Application
     {
         public static IServiceCollection AddWin2MqttApplication(this IServiceCollection services) =>
             services
-            .AddSingleton<ISensorFactory, SensorFactory>();
+            .AddSingleton<ISensorFactory, SensorFactory>()
+            .AddTransient<IIncomingMessagesProcessor, IncomingMessagesProcessor>();
     }
 }
